Fixed warnings.

+++ phoneProfiles/src/main/java/sk/henrichg/phoneprofiles/Profile.java
@@ -1407,7 +1407,7 @@ public class Profile {
         if (value < min) {
             return 0.0f;
         }
-        return (value - min) / (max - min);
+        return ((float)value - min) / (max - min);
     }
 
     // Hybrid Log Gamma constant values
